Added refractivity, fixed camera bug
The camera was coded in such a way that it could only work correctly from
the default position, as the pixel grid would stay in the same position in
space and trace rays from the camera position, through the grid, into the
scene.

The direction of rays are now computed with the same algorithm, but
without using the position of the camera. They are then traced from that
position with the computed direction.