Removed accidentally added and unimplemented files
The objects line and triangle were added during a class practice as a
result of a misunderstanding.

The class was covering 2D rasterization, instead of 3D raytracing, which
provided content that didn't apply to this assignment, so I left the
classes unimplemented.

CMake failed to refresh the file list, which allowed me to compile and
test the delivered assignment without noticing this problem, and I just
noticed it in class when I failed to show my results to the teacher.

I got 100/100 after solving it though, so happy ending!